+++ R/airdas_effort.R
@@ -5,8 +5,10 @@
 #' @param x \code{airdas_df} object; output from \code{\link{airdas_process}}, 
 #'  or a data frame that can be coerced to a \code{airdas_df} object
 #' @param method character; method to use to chop AirDAS data into effort segments
-#'   Can be "condition", "equallength", "section", or any partial match thereof (case sensitive)
-#' @param conditions character vector of names of conditions to include in segdata output.
+#'   Can be "condition", "equallength", "section", 
+#'   or any partial match thereof (case sensitive)
+#' @param conditions character vector of names of conditions 
+#'   to include in segdata output.
 #'   These values must be column names from the output of \code{\link{airdas_process}},
 #'   e.g. 'Bft', 'CCover', etc. The default is \code{NULL}, 
 #'   in which case all relevant conditions will be included.
@@ -20,6 +22,10 @@
 #'   Defaults to \code{NULL}, which uses one fewer than the number of cores
 #'   reported by \code{\link[parallel]{detectCores}}
 #'   Using 1 core likely will be faster for smaller datasets
+#' @param angle.min passed to \code{\link{airdas_sight}}
+#' @param bft.max numeric; the maximum Beaufort (column 'Bft') for which to 
+#'   mark a sighting as \code{TRUE} in 'included' (see Details). 
+#'   Default is 5.
 #' @param ... arguments passed to the chopping function specified using \code{method},
 #'   such as \code{seg.km} or \code{seg.min.km}
 #' 
@@ -64,12 +70,10 @@
 #'   which is used in \code{\link{airdas_effort_sight}} when summarizing
 #'   the number of sightings and animals for selected species.
 #'   \code{\link{airdas_effort_sight}} is a separate function to allow users to
-#'   personalize the included values as desired for their analysis.
+#'   personalize the 'included' values as desired for their specific analysis.
 #'   By default, i.e. in the output of this function, 'included' is \code{TRUE} if:
-#'   the sighting was made when on effort, 
-#'   by it was a standard sighting (see \code{\link{airdas_sight}}), 
-#'   in a Beaufort sea state less than or equal to five, 
-#'   and with a sighting angle less than or equal to 78.
+#'   the sighting was a standard sighting (see \code{\link{airdas_sight}}) 
+#'   and in a Beaufort sea state less than or equal to 'btf.max'.
 #' 
 #' @return List of three data frames: 
 #'   \itemize{
@@ -118,10 +122,12 @@ airdas_effort.data.frame <- function(x, ...) {
 
 #' @name airdas_effort
 #' @export
-airdas_effort.airdas_df <- function(x, method = c("condition", "equallength", "section"), 
-                                    conditions = NULL, 
-                                    distance.method = c("greatcircle", "lawofcosines", "haversine", "vincenty"),
-                                    num.cores = NULL, ...) {
+airdas_effort.airdas_df <- function(
+    x, method = c("condition", "equallength", "section"), 
+    conditions = NULL, 
+    distance.method = c("greatcircle", "lawofcosines", "haversine", "vincenty"),
+    num.cores = NULL, angle.min = 12, bft.max = 5, ...
+) {
   #----------------------------------------------------------------------------
   # Input checks
   
@@ -202,8 +208,8 @@ airdas_effort.airdas_df <- function(x, method = c("condition", "equallength", "s
   sightinfo <- x.eff %>% 
     left_join(select(segdata, .data$segnum, .data$mlat, .data$mlon), 
               by = "segnum") %>% 
-    airdas_sight() %>% 
-    mutate(included = .data$Bft <= 5 & abs(.data$Angle) <= 78 & .data$SightStd & .data$OnEffort, 
+    airdas_sight(angle.min = angle.min) %>% 
+    mutate(included = .data$Bft <= bft.max & .data$SightStd, 
            included = ifelse(is.na(.data$included), FALSE, .data$included)) %>% 
     select(-.data$dist_from_prev, -.data$cont_eff_section)