+++ tests/testthat/test-output.R
@@ -88,7 +88,8 @@ test_that("airdas_sight output has expected column names and classes", {
     "HKR", "Haze", "Kelp", "RedTide", "AltFt", "SpKnot",  
     "ObsL", "ObsB", "ObsR", "Rec", "VLI", "VLO", "VB", "VRI", "VRO", 
     "EffortDot", "EventNum", "file_das", "line_num", "file_type", 
-    "SightNo", "Obs", "Angle", "SightStd", "Mixed", "SpCode", "GsTotal", "GsSp", 
+    "SightNo", "Obs", "Angle", "ObsStd", "SightStd", 
+    "Mixed", "SpCode", "GsTotal", "GsSp", 
     "TurtleSize", "TurtleDirection", "TurtleTail"
   )
   
@@ -107,7 +108,8 @@ test_that("airdas_sight output has expected column names and classes with an ext
     "HKR", "Haze", "Kelp", "RedTide", "AltFt", "SpKnot",  
     "ObsL", "ObsB", "ObsR", "Rec", "VLI", "VLO", "VB", "VRI", "VRO", 
     "EffortDot", "EventNum", "file_das", "line_num", "file_type", "testrr", 
-    "SightNo", "Obs", "Angle", "SightStd", "Mixed", "SpCode", "GsTotal", "GsSp", 
+    "SightNo", "Obs", "Angle", "ObsStd", "SightStd", 
+    "Mixed", "SpCode", "GsTotal", "GsSp", 
     "TurtleSize", "TurtleDirection", "TurtleTail"
   )